refactor: colocate sys admin page components
