+++ frontend/config/webpack.entry.js
@@ -34,8 +34,8 @@ const entryFiles = {
   sharedFileViewVideo: '/shared-file-view-video.js',
   subscription: '/subscription.js',
   sysAdmin: '/sys-admin.js',
-  sysadminRepoHistory: '/pages/sys-admin/repo-history.js',
-  sysadminRepoSnapshot: '/pages/sys-admin/repo-snapshot.js',
+  sysadminRepoHistory: '/pages/sys-admin/repo-history/index.js',
+  sysadminRepoSnapshot: '/pages/sys-admin/repo-snapshot/index.js',
   TCAccept: '/tc-accept.js',
   TCView: '/tc-view.js',
   uploadLink: '/upload-link.js',
