+++ frontend/src/pages/sys-admin/index.js
@@ -10,46 +10,46 @@ import AdminLoginLogs from './admin-logs/login-logs';
 import AdminOperationLogs from './admin-logs/operation-logs';
 import DingtalkDepartments from './department-integration/dingtalk-departments';
 import WorkWeixinDepartments from './department-integration/work-weixin-departments';
-import Departments from './departments/departments';
+import Departments from './departments';
 import Devices from './devices';
 import DesktopDevices from './devices/desktop-devices';
 import DeviceErrors from './devices/devices-errors';
 import MobileDevices from './devices/mobile-devices';
 import FileScanRecords from './file-scan-records';
+import Groups from './groups';
 import GroupMembers from './groups/group-members';
 import GroupRepos from './groups/group-repos';
-import Groups from './groups/groups';
 import SearchGroups from './groups/search-groups';
 import Info from './info';
+import Institutions from './institutions';
 import InstitutionAdmins from './institutions/institution-admins';
 import InstitutionInfo from './institutions/institution-info';
 import InstitutionUsers from './institutions/institution-users';
-import Institutions from './institutions/institutions';
 import Invitations from './invitations/invitations';
 import LibrariesAndLinks from './libraries-and-links';
-import Logs from './logs-page';
-import FileAccessLogs from './logs-page/file-access-logs';
-import FIleTransferLogs from './logs-page/file-transfer-log';
-import FileUpdateLogs from './logs-page/file-update-logs';
-import GroupMemberAuditLogs from './logs-page/group-member-audit-logs';
-import LoginLogs from './logs-page/login-logs';
-import SharePermissionLogs from './logs-page/share-permission-logs';
 import MainPanel from './main-panel';
 import Notifications from './notifications/notifications';
+import Orgs from './orgs';
 import OrgAdmins from './orgs/org-admins';
 import OrgGroups from './orgs/org-groups';
 import OrgInfo from './orgs/org-info';
 import OrgRepos from './orgs/org-repos';
 import OrgUsers from './orgs/org-users';
-import Orgs from './orgs/orgs';
 import OrgsTrafficExceeded from './orgs/orgs-traffic-exceeded';
 import SearchOrgs from './orgs/search-orgs';
 import DirView from './repos/dir-view';
 import SearchRepos from './repos/search-repos';
 import WebSettings from './settings';
 import SidePanel from './side-panel';
-import StatisticLayout from './statistic/layout';
-import TermsAndConditions from './terms-and-conditions/terms-and-conditions';
+import StatisticLayout from './statistic';
+import Logs from './sys-logs';
+import FileAccessLogs from './sys-logs/file-access-logs';
+import FIleTransferLogs from './sys-logs/file-transfer-log';
+import FileUpdateLogs from './sys-logs/file-update-logs';
+import GroupMemberAuditLogs from './sys-logs/group-member-audit-logs';
+import LoginLogs from './sys-logs/login-logs';
+import SharePermissionLogs from './sys-logs/share-permission-logs';
+import TermsAndConditions from './terms-and-conditions';
 import { UsersLayout, UserLayout } from './users';
 import SearchUsers from './users/search-users';
 import UserGroups from './users/user-groups';
