+++ frontend/src/pages/sys-admin/libraries-and-links/index.js
@@ -1,19 +1,19 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import { Button } from 'reactstrap';
 import { useLocation, navigate, Router } from '@gatsbyjs/reach-router';
-import Icon from '../../components/icon';
-import toaster from '../../components/toast';
-import { gettext, siteRoot } from '../../utils/constants';
-import LinksNav from './links/links-nav';
-import ShareLinks from './links/share-links';
-import UploadLinks from './links/upload-links';
-import MainPanelTopbar from './main-panel-topbar';
-import AllRepos from './repos/all-repos';
-import AllWikis from './repos/all-wikis';
-import ReposNav from './repos/repos-nav';
-import SystemRepo from './repos/system-repo';
-import TrashRepos from './repos/trash-repos';
-import Search from './search';
+import Icon from '../../../components/icon';
+import toaster from '../../../components/toast';
+import { gettext, siteRoot } from '../../../utils/constants';
+import LinksNav from '../links/links-nav';
+import ShareLinks from '../links/share-links';
+import UploadLinks from '../links/upload-links';
+import MainPanelTopbar from '../main-panel-topbar';
+import AllRepos from '../repos/all-repos';
+import AllWikis from '../repos/all-wikis';
+import SystemRepo from '../repos/system-repo';
+import TrashRepos from '../repos/trash-repos';
+import Search from '../search';
+import ReposNav from './repos-nav';
 
 const LINKS_PATH_NAME_MAP = {
   'share-links': 'shareLinks',
