+++ frontend/src/pages/sys-admin/repo-snapshot/index.js
@@ -2,14 +2,14 @@ import React from 'react';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import { createRoot } from 'react-dom/client';
-import Icon from '../../components/icon';
-import Loading from '../../components/loading';
-import CommonToolbar from '../../components/toolbar/common-toolbar';
-import { gettext, siteRoot, mediaUrl, logoPath, logoWidth, logoHeight, siteTitle } from '../../utils/constants';
-import { systemAdminAPI } from '../../utils/system-admin-api';
-import { Utils } from '../../utils/utils';
+import Icon from '../../../components/icon';
+import Loading from '../../../components/loading';
+import CommonToolbar from '../../../components/toolbar/common-toolbar';
+import { gettext, siteRoot, mediaUrl, logoPath, logoWidth, logoHeight, siteTitle } from '../../../utils/constants';
+import { systemAdminAPI } from '../../../utils/system-admin-api';
+import { Utils } from '../../../utils/utils';
 
-import '../../css/repo-snapshot.css';
+import '../../../css/repo-snapshot.css';
 
 const {
   repoID, repoName,
