+++ frontend/src/pages/sys-admin/sys-logs/file-update-logs.js
@@ -1,27 +1,21 @@
-import React, { Component, Fragment } from 'react';
-import { navigate } from '@gatsbyjs/reach-router';
+import React, { Component } from 'react';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
 import PropTypes from 'prop-types';
+import CommitDetails from '../../../components/dialog/commit-details';
 import EmptyTip from '../../../components/empty-tip';
 import Loading from '../../../components/loading';
+import ModalPortal from '../../../components/modal-portal';
 import Paginator from '../../../components/paginator';
 import { gettext } from '../../../utils/constants';
 import { systemAdminAPI } from '../../../utils/system-admin-api';
 import { Utils } from '../../../utils/utils';
-import { LogRepoSelector, LogUserSelector } from '../log-selector';
 import UserLink from '../user-link';
+import { LogRepoSelector, LogUserSelector } from './log-selector';
 
 dayjs.extend(relativeTime);
 
 class Content extends Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      isItemFreezed: false
-    };
-  }
 
   getPreviousPage = () => {
     this.props.getLogsByPage(this.props.currentPage - 1);
@@ -31,37 +25,26 @@ class Content extends Component {
     this.props.getLogsByPage(this.props.currentPage + 1);
   };
 
-  toggleFreezeItem = (freezed) => {
-    this.setState({
-      isItemFreezed: freezed
-    });
-  };
-
   render() {
-    const {
-      loading, errorMsg, items,
-      perPage, currentPage, hasNextPage
-    } = this.props;
+    const { loading, errorMsg, items, perPage, currentPage, hasNextPage } = this.props;
     if (loading) {
       return <Loading />;
     } else if (errorMsg) {
       return <p className="error text-center">{errorMsg}</p>;
     } else {
       const emptyTip = (
-        <EmptyTip text={gettext('No file access logs')}>
+        <EmptyTip text={gettext('No file update logs')}>
         </EmptyTip>
       );
       const table = (
-        <Fragment>
-          <table>
+        <>
+          <table className="table-hover">
             <thead>
               <tr>
-                <th width="20%">{gettext('Name')}</th>
-                <th width="10%">{gettext('Type')}</th>
-                <th width="20%">{gettext('IP')}{' / '}{gettext('Device')}</th>
-                <th width="20%">{gettext('Date')}</th>
-                <th width="15%">{gettext('Library')}</th>
-                <th width="15%">{gettext('File')}{' / '}{gettext('Folder')}</th>
+                <th width="25%">{gettext('Name')}</th>
+                <th width="25%">{gettext('Date')}</th>
+                <th width="25%">{gettext('Library')}</th>
+                <th width="25%">{gettext('Action')}</th>
               </tr>
             </thead>
             {items &&
@@ -70,8 +53,6 @@ class Content extends Component {
                   return (<Item
                     key={index}
                     item={item}
-                    isFreezed={this.state.isItemFreezed}
-                    toggleFreezeItem={this.toggleFreezeItem}
                   />);
                 })}
               </tbody>
@@ -85,7 +66,7 @@ class Content extends Component {
             curPerPage={perPage}
             resetPerPage={this.props.resetPerPage}
           />
-        </Fragment>
+        </>
       );
       return items.length ? table : emptyTip;
     }
@@ -93,16 +74,15 @@ class Content extends Component {
 }
 
 Content.propTypes = {
-  loading: PropTypes.bool,
-  errorMsg: PropTypes.string,
-  items: PropTypes.array,
+  loading: PropTypes.bool.isRequired,
+  errorMsg: PropTypes.string.isRequired,
+  items: PropTypes.array.isRequired,
   getLogsByPage: PropTypes.func,
   resetPerPage: PropTypes.func,
   currentPage: PropTypes.number,
   perPage: PropTypes.number,
   pageInfo: PropTypes.object,
   hasNextPage: PropTypes.bool,
-  toggleFreezeItem: PropTypes.func
 };
 
 
@@ -111,63 +91,74 @@ class Item extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      isHighlighted: false,
+      isOpIconShown: false,
+      isCommitDetailsDialogOpen: false,
     };
   }
 
-  handleMouseEnter = () => {
-    if (!this.props.isFreezed) {
-      this.setState({
-        isHighlighted: true,
-      });
-    }
+  handleMouseOver = () => {
+    this.setState({
+      isOpIconShown: true
+    });
   };
 
-  handleMouseLeave = () => {
-    if (!this.props.isFreezed) {
-      this.setState({
-        isHighlighted: false,
-      });
-    }
+  handleMouseOut = () => {
+    this.setState({
+      isOpIconShown: false
+    });
   };
 
-  toggleFreezeItem = (freezed) => {
-    this.props.toggleFreezeItem(freezed);
-    if (!freezed) {
-      this.setState({
-        isHighlighted: false,
-      });
-    }
+
+  toggleCommitDetailsDialog = () => {
+    this.setState({
+      isCommitDetailsDialogOpen: !this.state.isCommitDetailsDialogOpen
+    });
+  };
+
+  showCommitDetails = (e) => {
+    e.preventDefault();
+    this.setState({
+      isCommitDetailsDialogOpen: !this.state.isCommitDetailsDialogOpen
+    });
   };
 
   render() {
-    const { isHighlighted } = this.state;
-    const { item } = this.props;
+    let { item } = this.props;
     return (
-      <tr className={isHighlighted ? 'tr-highlight' : ''} onMouseEnter={this.handleMouseEnter} onMouseLeave={this.handleMouseLeave}>
-        <td>
-          <UserLink email={item.email} name={item.name} />
-        </td>
-        <td>{item.event_type}</td>
-        <td>{item.ip}{' / '}{item.device || '--'}</td>
-        <td>{dayjs(item.time).fromNow()}</td>
-        <td>
-          {item.repo_name ? item.repo_name : gettext('Deleted')}
-        </td>
-        <td>{item.file_or_dir_name}</td>
-      </tr>
+      <>
+        <tr onMouseOver={this.handleMouseOver} onMouseOut={this.handleMouseOut}>
+          <td><UserLink email={item.email} name={item.name} /></td>
+          <td>{dayjs(item.time).fromNow()}</td>
+          <td>{item.repo_name ? item.repo_name : gettext('Deleted')}</td>
+          <td>
+            {item.file_operation}
+            {item.repo_name && !item.repo_encrypted &&
+              <a className="ml-1" href="#" onClick={this.showCommitDetails}>{gettext('Details')}</a>
+            }
+          </td>
+        </tr>
+        {this.state.isCommitDetailsDialogOpen &&
+          <ModalPortal>
+            <CommitDetails
+              repoID={item.repo_id}
+              commitID={item.commit_id}
+              commitTime={item.time}
+              toggleDialog={this.toggleCommitDetailsDialog}
+            />
+          </ModalPortal>
+        }
+      </>
     );
   }
 }
 
 
 Item.propTypes = {
-  item: PropTypes.object,
-  isFreezed: PropTypes.bool,
-  toggleFreezeItem: PropTypes.func,
+  item: PropTypes.object.isRequired,
 };
 
-class FileAccessLogs extends Component {
+
+class FileUpdateLogs extends Component {
 
   constructor(props) {
     super(props);
@@ -188,24 +179,28 @@ class FileAccessLogs extends Component {
     this.initPage = 1;
   }
 
+  toggleExportExcelDialog = () => {
+    this.setState({ isExportExcelDialogOpen: !this.state.isExportExcelDialogOpen });
+  };
+
   componentDidMount() {
     let urlParams = (new URL(window.location)).searchParams;
     const { currentPage, perPage } = this.state;
     this.setState({
       perPage: parseInt(urlParams.get('per_page') || perPage),
-      currentPage: parseInt(urlParams.get('page') || currentPage),
+      currentPage: parseInt(urlParams.get('page') || currentPage)
     }, () => {
       this.getLogsByPage(this.state.currentPage);
     });
   }
 
   getLogsByPage = (page) => {
-    const { perPage, selectedUsers, selectedRepos } = this.state;
+    let { perPage, selectedUsers, selectedRepos } = this.state;
     let emails = selectedUsers.map(user => user.email);
     let repos = selectedRepos.map(repo => repo.id);
-    systemAdminAPI.sysAdminListFileAccessLogs(page, perPage, { 'email': emails, 'repo': repos }).then((res) => {
+    systemAdminAPI.sysAdminListFileUpdateLogs(page, perPage, { 'email': emails, 'repo': repos }).then((res) => {
       this.setState({
-        logList: res.data.file_access_log_list,
+        logList: res.data.file_update_log_list,
         loading: false,
         currentPage: page,
         hasNextPage: res.data.has_next_page,
@@ -224,17 +219,6 @@ class FileAccessLogs extends Component {
     }, () => this.getLogsByPage(this.initPage));
   };
 
-  updateURL = (obj) => {
-    let url = new URL(location.href);
-    let searchParams = new URLSearchParams(url.search);
-    for (let key in obj) {
-      obj[key] == null ?
-        searchParams.delete(key) :
-        searchParams.set(key, obj[key]);
-    }
-    url.search = searchParams.toString();
-    navigate(url.toString());
-  };
 
   handleUserFilter = (user, shouldFetchData = true) => {
     const { selectedUsers } = this.state;
@@ -261,12 +245,25 @@ class FileAccessLogs extends Component {
     });
   };
 
+  handleSelectorToggle = (selectorType) => {
+    const { openSelector } = this.state;
+    const wasOpen = openSelector === selectorType;
+
+    this.setState({
+      openSelector: wasOpen ? null : selectorType
+    }, () => {
+      if (wasOpen) {
+        this.getLogsByPage(1);
+      }
+    });
+  };
+
   handleRepoFilter = (repo, shouldFetchData = true) => {
     const { selectedRepos } = this.state;
     let newSelectedRepos;
 
     if (repo === null) {
-      newSelectedRepos = [];
+      newSelectedRepos = selectedRepos;
     } else {
       const isSelected = selectedRepos.find(item => item.id === repo.id);
       if (isSelected) {
@@ -286,19 +283,6 @@ class FileAccessLogs extends Component {
     });
   };
 
-  handleSelectorToggle = (selectorType) => {
-    const { openSelector } = this.state;
-    const wasOpen = openSelector === selectorType;
-
-    this.setState({
-      openSelector: wasOpen ? null : selectorType
-    }, () => {
-      if (wasOpen) {
-        this.getLogsByPage(1);
-      }
-    });
-  };
-
   searchUsers = (value) => {
     return systemAdminAPI.sysAdminSearchUsers(value);
   };
@@ -308,15 +292,7 @@ class FileAccessLogs extends Component {
   };
 
   render() {
-    const {
-      logList,
-      currentPage, perPage, hasNextPage,
-      availableUsers,
-      selectedUsers,
-      availableRepos,
-      selectedRepos,
-      openSelector
-    } = this.state;
+    let { logList, currentPage, perPage, hasNextPage, availableUsers, selectedUsers, availableRepos, selectedRepos } = this.state;
     return (
       <div className="main-panel-center flex-row">
         <div className="cur-view-container">
@@ -327,7 +303,7 @@ class FileAccessLogs extends Component {
                 items={availableUsers}
                 selectedItems={selectedUsers}
                 onSelect={this.handleUserFilter}
-                isOpen={openSelector === 'user'}
+                isOpen={this.state.openSelector === 'user'}
                 onToggle={() => this.handleSelectorToggle('user')}
                 searchUsersFunc={this.searchUsers}
               />
@@ -335,7 +311,7 @@ class FileAccessLogs extends Component {
                 items={availableRepos}
                 selectedItems={selectedRepos}
                 onSelect={this.handleRepoFilter}
-                isOpen={openSelector === 'repo'}
+                isOpen={this.state.openSelector === 'repo'}
                 onToggle={() => this.handleSelectorToggle('repo')}
                 searchReposFunc={this.searchRepos}
               />
@@ -357,4 +333,4 @@ class FileAccessLogs extends Component {
   }
 }
 
-export default FileAccessLogs;
+export default FileUpdateLogs;
