+++ frontend/src/pages/sys-admin/sys-logs/log-selector/log-user-selector.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
 import { Input, Popover } from 'reactstrap';
 import PropTypes from 'prop-types';
-import Icon from '../../../components/icon';
-import SearchEmptyTip from '../../../components/search-empty-tip';
-import toaster from '../../../components/toast';
-import { gettext } from '../../../utils/constants';
-import { Utils } from '../../../utils/utils';
+import Icon from '../../../../components/icon';
+import SearchEmptyTip from '../../../../components/search-empty-tip';
+import toaster from '../../../../components/toast';
+import { gettext } from '../../../../utils/constants';
+import { Utils } from '../../../../utils/utils';
 
 import './index.css';
