+++ frontend/src/pages/sys-admin/terms-and-conditions/index.js
